Making possible to automatically build Addressables while building player

+++ Advanced/Play Asset Delivery/Assets/PlayAssetDelivery/Editor/PlayAssetDeliveryBuildProcessor.cs
@@ -15,10 +15,9 @@ namespace AddressablesPlayAssetDelivery.Editor
     /// <summary>
     /// Moves custom asset pack data from their default build location <see cref="BuildScriptPlayAssetDelivery"/> to their correct player build data location.
     /// For an Android App Bundle, bundles assigned to a custom asset pack must be located in their {asset pack name}.androidpack directory in the Assets folder.
-    /// The 'CustomAssetPacksData.json' file is also added to the built player's StreamingAssets file location.
     ///
     /// This script executes before the <see cref="AddressablesPlayerBuildProcessor"/> which moves all Addressables data to StreamingAssets.
-    public class PlayAssetDeliveryBuildProcessor : BuildPlayerProcessor, IPostGenerateGradleAndroidProject
+    public class PlayAssetDeliveryBuildProcessor : BuildPlayerProcessor
     {
         /// <summary>
         /// Returns the player build processor callback order.
@@ -34,18 +33,23 @@ namespace AddressablesPlayAssetDelivery.Editor
         /// <param name="buildPlayerContext">Contains data related to the player.</param>
         public override void PrepareForBuild(BuildPlayerContext buildPlayerContext)
         {
-            if (EditorUserBuildSettings.activeBuildTarget == BuildTarget.Android && EditorUserBuildSettings.buildAppBundle)
-                MoveDataForAppBundleBuild(buildPlayerContext);
+            if (EditorUserBuildSettings.activeBuildTarget != BuildTarget.Android)
+            {
+                return;
+            }
+            // need to check that Addressables are built for Android
+            if (EditorUserBuildSettings.buildAppBundle && (PlayerSettings.Android.splitApplicationBinary || TextureCompressionProcessor.EnabledTextureCompressionTargeting))
+            {
+                MoveDataForAppBundleBuild();
+            }
             else
+            {
                 MoveDataToDefaultLocation();
+            }
         }
 
-        static void MoveTextureCompressionData(BuildPlayerContext buildPlayerContext, string postfix)
+        static void MoveTextureCompressionData(string postfix)
         {
-            // need to check that data for specific texture compression exist (player settings might change)
-            var customAssetPackDataPath = Path.Combine($"{Addressables.StreamingAssetsSubFolder}{postfix}", CustomAssetPackUtility.kCustomAssetPackDataFilename);
-            buildPlayerContext.AddAdditionalPathToStreamingAssets(Path.Combine(CustomAssetPackUtility.BuildRootDirectory, customAssetPackDataPath), customAssetPackDataPath);
-
             var buildProcessorDataPath = Path.Combine(CustomAssetPackUtility.BuildRootDirectory, $"{Addressables.StreamingAssetsSubFolder}{postfix}", CustomAssetPackUtility.kBuildProcessorDataFilename);
             if (File.Exists(buildProcessorDataPath))
             {
@@ -57,6 +61,10 @@ namespace AddressablesPlayAssetDelivery.Editor
                     string assetsFolderPath = Path.Combine(CustomAssetPackUtility.PackContentRootDirectory, entry.AssetsSubfolderPath);
                     if (File.Exists(entry.BundleBuildPath))
                     {
+                        if (File.Exists(assetsFolderPath))
+                        {
+                            File.Delete(assetsFolderPath);
+                        }
                         if (TextureCompressionProcessor.EnabledTextureCompressionTargeting && string.IsNullOrEmpty(postfix))
                         {
                             File.Copy(entry.BundleBuildPath, assetsFolderPath);
@@ -80,26 +88,19 @@ namespace AddressablesPlayAssetDelivery.Editor
         /// <summary>
         /// Move custom asset pack data from their build location to their App Bundle data location.
         /// </summary>
-        /// <param name="buildPlayerContext">Contains data related to the player.</param>
-        public static void MoveDataForAppBundleBuild(BuildPlayerContext buildPlayerContext)
+        public static void MoveDataForAppBundleBuild()
         {
             try
             {
                 AssetDatabase.StartAssetEditing();
-
-                MoveTextureCompressionData(buildPlayerContext, "");
-
+                MoveTextureCompressionData("");
                 if (TextureCompressionProcessor.EnabledTextureCompressionTargeting)
                 {
                     foreach (var textureCompression in PlayerSettings.Android.textureCompressionFormats)
                     {
-                        var postfixDst = TextureCompressionProcessor.TcfPostfix(textureCompression);
-                        var postfixSrc = textureCompression == PlayerSettings.Android.textureCompressionFormats[0] ? "" : postfixDst;
-                        MoveTextureCompressionData(buildPlayerContext, postfixDst);
-                        buildPlayerContext.AddAdditionalPathToStreamingAssets($"{Addressables.BuildPath}{postfixSrc}", $"{Addressables.StreamingAssetsSubFolder}{postfixDst}");
+                        MoveTextureCompressionData(TextureCompressionProcessor.TcfPostfix(textureCompression));
                     }
                 }
-                buildPlayerContext.AddAdditionalPathToStreamingAssets(Addressables.BuildPath, Addressables.StreamingAssetsSubFolder);
             }
             catch (Exception e)
             {
@@ -114,13 +115,11 @@ namespace AddressablesPlayAssetDelivery.Editor
         /// <summary>
         /// Move custom asset pack data from their App Bundle data location to to their build location.
         /// </summary>
-        // need to check this method
         public static void MoveDataToDefaultLocation()
         {
             try
             {
                 AssetDatabase.StartAssetEditing();
-
                 if (File.Exists(CustomAssetPackUtility.BuildProcessorDataPath))
                 {
                     string contents = File.ReadAllText(CustomAssetPackUtility.BuildProcessorDataPath);
@@ -129,8 +128,12 @@ namespace AddressablesPlayAssetDelivery.Editor
                     foreach (BuildProcessorDataEntry entry in data.Entries)
                     {
                         string assetsFolderPath = Path.Combine(CustomAssetPackUtility.PackContentRootDirectory, entry.AssetsSubfolderPath);
-                        if (File.Exists(assetsFolderPath))
+                        if (File.Exists(assetsFolderPath) && Directory.Exists(Path.GetDirectoryName(entry.BundleBuildPath)))
                         {
+                            if (File.Exists(entry.BundleBuildPath))
+                            {
+                                File.Delete(entry.BundleBuildPath);
+                            }
                             string metaFilePath = AssetDatabase.GetTextMetaFilePathFromAssetPath(assetsFolderPath);
                             File.Move(assetsFolderPath, entry.BundleBuildPath);
                             File.Delete(metaFilePath);
@@ -147,7 +150,54 @@ namespace AddressablesPlayAssetDelivery.Editor
                 AssetDatabase.StopAssetEditing();
             }
         }
+    }
 
+    /// <summary>
+    /// Moves all install-time Addressables data and Addressables json files to StreamingAssets.
+    /// This script executes after the <see cref="AddressablesPlayerBuildProcessor"/> which moves other Addressables data to StreamingAssets.
+    public class PlayAssetDeliverySecondBuildProcessor : BuildPlayerProcessor, IPostGenerateGradleAndroidProject
+    {
+        /// <summary>
+        /// Returns the player build processor callback order.
+        /// </summary>
+        public override int callbackOrder
+        {
+            get { return 2; }
+        }
+
+        /// <summary>
+        /// Invoked before performing a Player build. Moves AssetBundles to their correct data location based on the build target platform.
+        /// </summary>
+        /// <param name="buildPlayerContext">Contains data related to the player.</param>
+        public override void PrepareForBuild(BuildPlayerContext buildPlayerContext)
+        {
+            if (EditorUserBuildSettings.activeBuildTarget == BuildTarget.Android && EditorUserBuildSettings.buildAppBundle && 
+                (PlayerSettings.Android.splitApplicationBinary || TextureCompressionProcessor.EnabledTextureCompressionTargeting))
+            {
+                MoveInstallTimeDataToStreamingAssets(buildPlayerContext);
+            }
+        }
+
+        static void MoveInstallTimeDataToStreamingAssets(BuildPlayerContext buildPlayerContext)
+        {
+            var customAssetPackDataPath = Path.Combine(Addressables.StreamingAssetsSubFolder, CustomAssetPackUtility.kCustomAssetPackDataFilename);
+            buildPlayerContext.AddAdditionalPathToStreamingAssets(Path.Combine(CustomAssetPackUtility.BuildRootDirectory, customAssetPackDataPath), customAssetPackDataPath);
+
+            if (TextureCompressionProcessor.EnabledTextureCompressionTargeting)
+            {
+                foreach (var textureCompression in PlayerSettings.Android.textureCompressionFormats)
+                {
+                    var postfixDst = TextureCompressionProcessor.TcfPostfix(textureCompression);
+                    var postfixSrc = textureCompression == PlayerSettings.Android.textureCompressionFormats[0] ? "" : postfixDst;
+                    // need to check that data for specific texture compression exist (player settings might change)
+                    customAssetPackDataPath = Path.Combine($"{Addressables.StreamingAssetsSubFolder}{postfixDst}", CustomAssetPackUtility.kCustomAssetPackDataFilename);
+                    buildPlayerContext.AddAdditionalPathToStreamingAssets(Path.Combine(CustomAssetPackUtility.BuildRootDirectory, customAssetPackDataPath), customAssetPackDataPath);
+                    buildPlayerContext.AddAdditionalPathToStreamingAssets($"{Addressables.BuildPath}{postfixSrc}", $"{Addressables.StreamingAssetsSubFolder}{postfixDst}");
+                }
+            }
+        }
+
+        // probably create separate class here
         const string kUnityAssetPackTextureCompressions = "UnityTextureCompressionsAssetPack";
         const string kUnityAssetPackStreamingAssets = "UnityStreamingAssetsPack";
 
@@ -160,13 +210,35 @@ namespace AddressablesPlayAssetDelivery.Editor
             Directory.Move($"{fromPath}{postfix}", $"{toPath}{postfix}");
         }
 
+        void CheckStreamingAssetsPack(string gradleProjectPath)
+        {
+            // Check if UnityAssetPackStreamingAssets is empty after moving data from it and if it is, then delete its mentions from gradle files
+            var streamingAssetPack = Path.Combine(gradleProjectPath, kUnityAssetPackStreamingAssets);
+            foreach (var file in Directory.GetFiles(streamingAssetPack, "*", SearchOption.AllDirectories))
+            {
+                if (!file.EndsWith("build.gradle") && !Path.GetRelativePath(streamingAssetPack, file).StartsWith("build"))
+                {
+                    return;
+                }
+            }
+            Directory.Delete(Path.Combine(gradleProjectPath, kUnityAssetPackStreamingAssets), true);
+            var settingsGradle = File.ReadAllText(Path.Combine(gradleProjectPath, "settings.gradle"));
+            File.WriteAllText(Path.Combine(gradleProjectPath, "settings.gradle"), settingsGradle.Replace("include ':UnityStreamingAssetsPack'", ""));
+            var launcherGradle = File.ReadAllText(Path.Combine(gradleProjectPath, "launcher/build.gradle"));
+            File.WriteAllText(Path.Combine(gradleProjectPath, "launcher/build.gradle"), launcherGradle.Replace(", \":UnityStreamingAssetsPack\"", ""));
+        }
+
         public void OnPostGenerateGradleAndroidProject(string path)
         {
+            if (!EditorUserBuildSettings.buildAppBundle)
+            {
+                return;
+            }
+            // We get path to unityLibrary, move above
+            var gradleProjectPath = Path.GetFullPath(Path.Combine(path, ".."));
+            var assetPacksUpdated = false;
             if (TextureCompressionProcessor.EnabledTextureCompressionTargeting)
             {
-                // We get path to unityLibrary, move above
-                var gradleProjectPath = Path.GetFullPath(Path.Combine(path, ".."));
-
                 // Move all install-time addressables data to UnityAssetPackTextureCompressions as UnityAssetPackStreamingAssets might be not install-time
                 var addressablesStreamingResourcesPath = Path.Combine(gradleProjectPath, kUnityAssetPackStreamingAssets, CustomAssetPackUtility.CustomAssetPacksAssetsPath);
                 var addressablesResourcesPath = Path.Combine(gradleProjectPath, kUnityAssetPackTextureCompressions, CustomAssetPackUtility.CustomAssetPacksAssetsPath);
@@ -177,18 +249,20 @@ namespace AddressablesPlayAssetDelivery.Editor
                     MoveAddressablesData(addressablesStreamingResourcesPath, addressablesResourcesPath, TextureCompressionProcessor.TcfPostfix(textureCompression));
                 }
 
-                // Check if UnityAssetPackStreamingAssets is empty after moving data from it and if it is, then delete its mentions from gradle files
-                var remain = Directory.GetFiles(Path.Combine(gradleProjectPath, kUnityAssetPackStreamingAssets), "*", SearchOption.AllDirectories);
-                if (remain.Length == 1 && remain[0].EndsWith("build.gradle"))
-                {
-                    Directory.Delete(Path.Combine(gradleProjectPath, kUnityAssetPackStreamingAssets), true);
-                    var settingsGradle = File.ReadAllText(Path.Combine(gradleProjectPath, "settings.gradle"));
-                    File.WriteAllText(Path.Combine(gradleProjectPath, "settings.gradle"), settingsGradle.Replace("include ':UnityStreamingAssetsPack'", ""));
-                    var launcherGradle = File.ReadAllText(Path.Combine(gradleProjectPath, "launcher/build.gradle"));
-                    File.WriteAllText(Path.Combine(gradleProjectPath, "launcher/build.gradle"), launcherGradle.Replace(", \":UnityStreamingAssetsPack\"", ""));
-                }
+                CheckStreamingAssetsPack(gradleProjectPath);
+                assetPacksUpdated = true;
             }
-            // Probably recheck asset pack sizes here
+            else if (PlayerSettings.Android.splitApplicationBinary)
+            {
+                // Move all install-time addressables data to unityLibrary from UnityAssetPackStreamingAssets
+                // If there is no UnityAssetPackStreamingAssets, this means that all data are packed to UnityDataAssetPack
+                // and it's must be install-time (need to check)
+                assetPacksUpdated = true;
+            }
+            if (assetPacksUpdated)
+            {
+                // Recheck asset pack sizes here
+            }
         }
     }
 }
